% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_features}
\alias{data_features}
\title{Simulated Dataframe}
\format{
A data frame with 100 rows and 101 variables:
}
\usage{
data(data_features)
}
\description{
A simulated dataframe with 100 rows and 101 columns. Col1 is an ID column.
The remaining columns are features with randomly assigned values.
Once values were assigned, 5\% missingness was introduced for 50 random columns
and 50 random rows.
}
\keyword{datasets}
